Moved tier visibility usage from portal settings to tier object
refs https://github.com/TryGhost/Team/issues/1387

We have moved away from the portal_products and portal_plans settings to using the visibility property on tiers to determine whether or not a tier should be visible in Portal. This updates Portal to use new tier visibility property to determine visibility at base API level, and in future we will update to remove all usage of portal settings.